Refactor whisper-cli to be more modular
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/LeopoldFortunatus/whisper-cli?shareId=XXXX-XXXX-XXXX-XXXX).
